fixup! [TSAN] Add __tsan_check_no_mutexes_held helper
- "locked" -> "held". Rename functions, enum value, and test cases to match
- init mutex and don't use threads in tests
- undo mutex acquired/created change
- fix comment on interface function
