+++ compiler-rt/include/sanitizer/tsan_interface.h
@@ -126,8 +126,8 @@ void __tsan_mutex_post_signal(void *addr, unsigned flags);
 void __tsan_mutex_pre_divert(void *addr, unsigned flags);
 void __tsan_mutex_post_divert(void *addr, unsigned flags);
 
-// Annotate that no mutexes can be held. If we are holding mutexes, then
-// TSan will print a bug report.
+// Check that the current thread does not hold any mutexes,
+// report a bug report otherwise.
 void __tsan_check_no_mutexes_held();
 
 // External race detection API.
